Transformed the case diary entry text area in a div instead of a text box in order to edit the text, now the user is able to change text color and make it bold, italic and etc.
Added readme file.

CS Files area were rebuilt. Now it is a button on the body of the case and the cs file registering is made directly to the pop up for easy verification history. Now the user is able to edit and delete cs file entries.

Fixed refresh pending tasks on modal open. The pending tasks list was only being rendered when the main view was updated. This caused the list in the "Pending Tasks" modal to become stale if tasks were modified after the initial render. This change updates the "Tarefas" button event listener to call the renderPendingTasks() function just before showing the modal. This ensures the list is always up-to-date with the current state of pending tasks. Now the pending tasks have different colors for an easier identification.