Added functionality to hit Enter Key for Adding Tags and Tasks.
Added Color-Coded Links in Pending Tasks Modal.

The renderPostLiveList function has been updated to filter out any case that has the isArchived flag set to true. This ensures that the "Post-live" list only displays active, non-archived post-live cases, providing a more accurate view.

The "Archive Case" button now dynamically changes to a "Reactivate Case" button when viewing an archived case. Clicking this button sets the case's isArchived status to false, effectively moving it from the "Archived" list back to the main active cases list (or the "Post-live" list if applicable).

Added a counter to the "Tarefas" (Tasks) button, providing an at-a-glance view of the total number of pending tasks across all cases. The count is displayed in parentheses next to the button text and is dynamically updated whenever a task is: Added, Marked as complete or Deleted.

Added a new search feature to the application, allowing users to search for cases based on various criteria. A"Search" button has been added to the main header, providing an entry point to the search functionality.